Improvements to throw interaction
Instead of always using an angle of 45 degrees, this implements a variable angle which depends on the angle between the camera and the table. This makes the throw interaction more intuitive. Also, you can now control the camera in the test scene by holding the right mouse button and moving the mouse.